refactoring: split Game from Board, add utils.py module
